Remove predictor related files and configurations
All predictor-related files and configurations have been deleted, including the main application, configs, tests, and resources. These deletions affect application execution and the build process. Always check for dependencies before merging this change into any active branches.
