+++ deploy.sh
@@ -378,7 +378,7 @@ build_docker() {
 
   components=("opencdx/mongodb" "opencdx/admin" "opencdx/config" "opencdx/tinkar"
     "opencdx/audit" "opencdx/communications" "opencdx/media" "opencdx/health" "opencdx/connected-lab"
-    "opencdx/iam" "opencdx/routine" "opencdx/protector" "opencdx/predictor"
+    "opencdx/iam" "opencdx/routine" "opencdx/protector"
     "opencdx/questionnaire" "opencdx/classification" "opencdx/gateway" "opencdx/logistics"
     "opencdx/discovery" "opencdx/anf" "opencdx/graphql")
