+++ opencdx-admin/src/main/resources/public/javadoc/cdx/opencdx/predictor/controller/OpenCDXGrpcPredictorController.html
@@ -86,7 +86,7 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance">cdx.opencdx.grpc.neural.predictor.NeuralPredictorServiceGrpc.NeuralPredictorServiceImplBase
-<div class="inheritance">cdx.opencdx.predictor.controller.OpenCDXGrpcPredictorController</div>
+<div class="inheritance">cdx.opencdx.classification.controller.OpenCDXGrpcPredictorController</div>
 </div>
 </div>
 <section class="class-description" id="class-description">
@@ -113,7 +113,7 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Constructor</div>
 <div class="table-header col-last">Description</div>
-<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(cdx.opencdx.predictor.service.OpenCDXPredictorService)" class="member-name-link">OpenCDXGrpcPredictorController</a><wbr>(<a href="../service/OpenCDXPredictorService.html" title="interface in cdx.opencdx.predictor.service">OpenCDXPredictorService</a>&nbsp;predictorService)</code></div>
+<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(cdx.opencdx.classification.service.OpenCDXPredictorService)" class="member-name-link">OpenCDXGrpcPredictorController</a><wbr>(<a href="../service/OpenCDXPredictorService.html" title="interface in cdx.opencdx.predictor.service">OpenCDXPredictorService</a>&nbsp;predictorService)</code></div>
 <div class="col-last even-row-color">
 <div class="block">Constructor using the predictorService.</div>
 </div>
@@ -158,7 +158,7 @@ loadScripts(document, 'script');</script>
 <h2>Constructor Details</h2>
 <ul class="member-list">
 <li>
-<section class="detail" id="&lt;init&gt;(cdx.opencdx.predictor.service.OpenCDXPredictorService)">
+<section class="detail" id="&lt;init&gt;(cdx.opencdx.classification.service.OpenCDXPredictorService)">
 <h3>OpenCDXGrpcPredictorController</h3>
 <div class="member-signature"><span class="annotations"><a href="https://docs.spring.io/spring-framework/docs/3.2.1.RELEASE/javadoc-api/org/springframework/beans/factory/annotation/Autowired.html" title="class or interface in org.springframework.beans.factory.annotation" class="external-link">@Autowired</a>
 </span><span class="modifiers">public</span>&nbsp;<span class="element-name">OpenCDXGrpcPredictorController</span><wbr><span class="parameters">(<a href="../service/OpenCDXPredictorService.html" title="interface in cdx.opencdx.predictor.service">OpenCDXPredictorService</a>&nbsp;predictorService)</span></div>
