+++ opencdx-admin/src/main/resources/public/javadoc/cdx/opencdx/predictor/controller/OpenCDXRestPredictorController.html
@@ -85,7 +85,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class OpenCDXRestPredictorController" class="title">Class OpenCDXRestPredictorController</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">cdx.opencdx.predictor.controller.OpenCDXRestPredictorController</div>
+<div class="inheritance">cdx.opencdx.classification.controller.OpenCDXRestPredictorController</div>
 </div>
 <section class="class-description" id="class-description">
 <hr>
@@ -108,7 +108,7 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Constructor</div>
 <div class="table-header col-last">Description</div>
-<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(cdx.opencdx.predictor.service.OpenCDXPredictorService)" class="member-name-link">OpenCDXRestPredictorController</a><wbr>(<a href="../service/OpenCDXPredictorService.html" title="interface in cdx.opencdx.predictor.service">OpenCDXPredictorService</a>&nbsp;predictorService)</code></div>
+<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(cdx.opencdx.classification.service.OpenCDXPredictorService)" class="member-name-link">OpenCDXRestPredictorController</a><wbr>(<a href="../service/OpenCDXPredictorService.html" title="interface in cdx.opencdx.predictor.service">OpenCDXPredictorService</a>&nbsp;predictorService)</code></div>
 <div class="col-last even-row-color">
 <div class="block">Constructor that takes a predictorService.</div>
 </div>
@@ -149,7 +149,7 @@ loadScripts(document, 'script');</script>
 <h2>Constructor Details</h2>
 <ul class="member-list">
 <li>
-<section class="detail" id="&lt;init&gt;(cdx.opencdx.predictor.service.OpenCDXPredictorService)">
+<section class="detail" id="&lt;init&gt;(cdx.opencdx.classification.service.OpenCDXPredictorService)">
 <h3>OpenCDXRestPredictorController</h3>
 <div class="member-signature"><span class="annotations"><a href="https://docs.spring.io/spring-framework/docs/3.2.1.RELEASE/javadoc-api/org/springframework/beans/factory/annotation/Autowired.html" title="class or interface in org.springframework.beans.factory.annotation" class="external-link">@Autowired</a>
 </span><span class="modifiers">public</span>&nbsp;<span class="element-name">OpenCDXRestPredictorController</span><wbr><span class="parameters">(<a href="../service/OpenCDXPredictorService.html" title="interface in cdx.opencdx.predictor.service">OpenCDXPredictorService</a>&nbsp;predictorService)</span></div>
