+++ opencdx-admin/src/main/resources/public/javadoc/cdx/opencdx/predictor/service/impl/OpenCDXPredictorServiceImpl.html
@@ -85,7 +85,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class OpenCDXPredictorServiceImpl" class="title">Class OpenCDXPredictorServiceImpl</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">cdx.opencdx.predictor.service.impl.OpenCDXPredictorServiceImpl</div>
+<div class="inheritance">cdx.opencdx.classification.service.impl.OpenCDXPredictorServiceImpl</div>
 </div>
 <section class="class-description" id="class-description">
 <dl class="notes">
