+++ opencdx-admin/src/main/resources/public/javadoc/index-all.html
@@ -5537,7 +5537,7 @@ loadScripts(document, 'script');</script>
 <dd>
 <div class="block">gRPC Controller for Predictor World.</div>
 </dd>
-<dt><a href="cdx/opencdx/predictor/controller/OpenCDXGrpcPredictorController.html#%3Cinit%3E(cdx.opencdx.predictor.service.OpenCDXPredictorService)" class="member-name-link">OpenCDXGrpcPredictorController(OpenCDXPredictorService)</a> - Constructor for class cdx.opencdx.predictor.controller.<a href="cdx/opencdx/predictor/controller/OpenCDXGrpcPredictorController.html" title="class in cdx.opencdx.predictor.controller">OpenCDXGrpcPredictorController</a></dt>
+<dt><a href="cdx/opencdx/predictor/controller/OpenCDXGrpcPredictorController.html#%3Cinit%3E(cdx.opencdx.classification.service.OpenCDXPredictorService)" class="member-name-link">OpenCDXGrpcPredictorController(OpenCDXPredictorService)</a> - Constructor for class cdx.opencdx.predictor.controller.<a href="cdx/opencdx/predictor/controller/OpenCDXGrpcPredictorController.html" title="class in cdx.opencdx.predictor.controller">OpenCDXGrpcPredictorController</a></dt>
 <dd>
 <div class="block">Constructor using the predictorService.</div>
 </dd>
@@ -6668,7 +6668,7 @@ loadScripts(document, 'script');</script>
 <dd>
 <div class="block">Controller for the /predictor API's.</div>
 </dd>
-<dt><a href="cdx/opencdx/predictor/controller/OpenCDXRestPredictorController.html#%3Cinit%3E(cdx.opencdx.predictor.service.OpenCDXPredictorService)" class="member-name-link">OpenCDXRestPredictorController(OpenCDXPredictorService)</a> - Constructor for class cdx.opencdx.predictor.controller.<a href="cdx/opencdx/predictor/controller/OpenCDXRestPredictorController.html" title="class in cdx.opencdx.predictor.controller">OpenCDXRestPredictorController</a></dt>
+<dt><a href="cdx/opencdx/predictor/controller/OpenCDXRestPredictorController.html#%3Cinit%3E(cdx.opencdx.classification.service.OpenCDXPredictorService)" class="member-name-link">OpenCDXRestPredictorController(OpenCDXPredictorService)</a> - Constructor for class cdx.opencdx.predictor.controller.<a href="cdx/opencdx/predictor/controller/OpenCDXRestPredictorController.html" title="class in cdx.opencdx.predictor.controller">OpenCDXRestPredictorController</a></dt>
 <dd>
 <div class="block">Constructor that takes a predictorService.</div>
 </dd>
