+++ opencdx-classification/src/main/java/cdx/opencdx/classification/controller/OpenCDXGrpcPredictorController.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cdx.opencdx.predictor.controller;
+package cdx.opencdx.classification.controller;
 
+import cdx.opencdx.classification.service.OpenCDXPredictorService;
 import cdx.opencdx.grpc.neural.predictor.NeuralPredictorServiceGrpc;
 import cdx.opencdx.grpc.neural.predictor.PredictorRequest;
 import cdx.opencdx.grpc.neural.predictor.PredictorResponse;
-import cdx.opencdx.predictor.service.OpenCDXPredictorService;
 import io.grpc.stub.StreamObserver;
 import io.micrometer.observation.annotation.Observed;
 import lombok.extern.slf4j.Slf4j;
