+++ opencdx-classification/src/main/java/cdx/opencdx/classification/controller/OpenCDXRestPredictorController.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cdx.opencdx.predictor.controller;
+package cdx.opencdx.classification.controller;
 
+import cdx.opencdx.classification.service.OpenCDXPredictorService;
 import cdx.opencdx.grpc.neural.predictor.PredictorRequest;
 import cdx.opencdx.grpc.neural.predictor.PredictorResponse;
-import cdx.opencdx.predictor.service.OpenCDXPredictorService;
 import io.micrometer.observation.annotation.Observed;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
