+++ opencdx-classification/src/main/java/cdx/opencdx/classification/service/OpenCDXPredictorService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cdx.opencdx.predictor.service;
+package cdx.opencdx.classification.service;
 
 import cdx.opencdx.grpc.neural.predictor.PredictorRequest;
 import cdx.opencdx.grpc.neural.predictor.PredictorResponse;
