+++ opencdx-classification/src/main/java/cdx/opencdx/classification/service/impl/OpenCDXPredictorServiceImpl.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cdx.opencdx.predictor.service.impl;
+package cdx.opencdx.classification.service.impl;
 
+import cdx.opencdx.classification.service.OpenCDXPredictorService;
 import cdx.opencdx.commons.data.OpenCDXIdentifier;
 import cdx.opencdx.commons.service.OpenCDXDocumentValidator;
 import cdx.opencdx.grpc.neural.predictor.PredictorOutput;
 import cdx.opencdx.grpc.neural.predictor.PredictorRequest;
 import cdx.opencdx.grpc.neural.predictor.PredictorResponse;
-import cdx.opencdx.predictor.service.OpenCDXPredictorService;
 import io.micrometer.observation.annotation.Observed;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
