--- opencdx-predictor/build.gradle
@@ -1,74 +1,0 @@
-/*
- * This file was generated by the Gradle "init" task.
- */
-plugins {
-    id("application")
-}
-
-dependencies {
-
-    // Spring
-    implementation "org.springframework.boot:spring-boot-starter-actuator:${springBootVersion}"
-    implementation "org.springframework.boot:spring-boot-starter-web:${springBootVersion}"
-    implementation "org.springframework.boot:spring-boot-starter-validation:${springBootVersion}"
-    implementation "org.springframework.boot:spring-boot-starter-test:${springBootVersion}"
-    implementation "org.springframework.boot:spring-boot-starter-data-mongodb:${springBootVersion}"
-    implementation "org.springframework.boot:spring-boot-starter-aop:${springBootVersion}"
-    implementation "io.github.lognet:grpc-spring-boot-starter:${grpcSpringBootStarterVersion}"
-    implementation "org.springframework.boot:spring-boot-starter-security:${springBootVersion}"
-    implementation "org.springframework.boot:spring-boot-starter-cache:${springBootVersion}"
-    
-
-    implementation "org.springframework.cloud:spring-cloud-starter-config"
-    implementation "org.springframework.cloud:spring-cloud-starter-netflix-eureka-client"
-
-    //Database
-    implementation "com.github.cloudyrock.mongock:mongock-spring-v5:${mongoData}"
-    implementation "com.github.cloudyrock.mongock:mongodb-springdata-v3-driver:${mongoData}"
-    implementation "com.github.cloudyrock.mongock:mongock-bom:${mongoData}"
-
-    // Swagger
-    implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:${springDocVersion}"
-
-    //GRPC - To enable the GRPC uncomment the following two lines
-    implementation "com.google.protobuf:protobuf-java:${protobufVersion}"
-    implementation project(":opencdx-proto")
-    implementation project(":opencdx-commons")
-    implementation project(":opencdx-client")
-    implementation 'io.micrometer:micrometer-observation'
-    implementation 'io.micrometer:micrometer-tracing'
-    implementation 'io.micrometer:micrometer-tracing-bridge-brave'
-    runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
-    implementation 'io.zipkin.reporter2:zipkin-reporter-brave'
-    runtimeOnly "com.github.loki4j:loki-logback-appender:${lokiLogBackAppenderVersion}"
-    testImplementation "io.nats:nats-spring:${natsSpringVersion}"
-}
-
-sourceSets {
-    main {
-        java {
-            srcDirs = ["src/main/java", "build/generated/sources/annotationProcessor/java/main"]
-        }
-        resources {
-            srcDirs = ["src/main/resources"]
-        }
-
-    }
-    test {
-        java {
-            srcDirs = ["src/test/java", "build/generated/sources/annotationProcessor/java/test"]
-        }
-        resources {
-            srcDirs = ["src/test/resources"]
-        }
-    }
-}
-
-bootJar {
-    enabled = true
-}
-
-application {
-    // Define the main class for the application.
-    mainClass = "cdx.opencdx.predictor.Application"
-}
