--- opencdx-predictor/lombok.config
@@ -1,2 +1,0 @@
-config.stopBubbling = true
-lombok.addLombokGeneratedAnnotation = true
