--- opencdx-predictor/src/main/java/cdx/opencdx/predictor/config/AppProperties.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2024 Safe Health Systems, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package cdx.opencdx.predictor.config;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-/**
- * Applicaiton Properties
- */
-@Data
-@ConfigurationProperties("opencdx.predictor")
-public class AppProperties {
-    /**
-     * Default Constructor
-     */
-    public AppProperties() {
-        // Explicit declaration to prevent this class from inadvertently being made instantiable
-    }
-
-    private String format;
-}
