--- opencdx-predictor/src/main/java/cdx/opencdx/predictor/config/package-info.java
@@ -1,12 +1,0 @@
-/**
- * This is the principal configuration file for Spring.
- * <p>
- * All Spring configuration files should be kept in this package. The only exception pertains to security files,
- * which are maintained in a separate package.
- * </p>
- *
- * @author Safe Health LLC.
- * @version 1.0
- * @since 1.0
- */
-package cdx.opencdx.predictor.config;
