--- opencdx-predictor/src/main/java/cdx/opencdx/predictor/controller/package-info.java
@@ -1,11 +1,0 @@
-/**
- * This package is dedicated to the maintenance of the REST and gRPC API controllers.
- * <p>
- * The controller files, which constitute the Rest API for this application, are maintained here.
- * </p>
- *
- * @author Safe Health LLC.
- * @version 1.0
- * @since 1.0
- */
-package cdx.opencdx.predictor.controller;
