--- opencdx-predictor/src/main/java/cdx/opencdx/predictor/package-info.java
@@ -1,8 +1,0 @@
-/**
- * This system is responsible for delivering predictions. Based on the given set of inputs, it analyzes patterns, computes correspondences and subsequently generates the most probable outcomes. Ideal for forecasting, statistical analysis, and other predictive modeling applications.
- *
- * @author Safe Health LLC.
- * @version 1.0
- * @since 1.0
- */
-package cdx.opencdx.predictor;
