--- opencdx-predictor/src/main/java/cdx/opencdx/predictor/security/SecurityConfig.java
@@ -1,94 +1,0 @@
-/*
- * Copyright 2024 Safe Health Systems, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package cdx.opencdx.predictor.security;
-
-import cdx.opencdx.commons.annotations.ExcludeFromJacocoGeneratedReport;
-import cdx.opencdx.commons.security.JwtTokenFilter;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfigurationSource;
-
-/**
- * Application Level Security Configuration
- */
-@Slf4j
-@Configuration
-@EnableWebSecurity
-@ExcludeFromJacocoGeneratedReport
-public class SecurityConfig {
-
-    private final AuthenticationConfiguration configuration;
-    private final JwtTokenFilter jwtTokenFilter;
-    private final CorsConfigurationSource corsConfigurationSource;
-    /**
-     * Constructore for Security Configuraiton
-     * @param configuration Authentication Configuration to use
-     * @param jwtTokenFilter Filter for using JWT
-     * @param corsConfigurationSource CORS configuration
-     */
-    public SecurityConfig(
-            AuthenticationConfiguration configuration,
-            JwtTokenFilter jwtTokenFilter,
-            CorsConfigurationSource corsConfigurationSource) {
-        this.configuration = configuration;
-        this.jwtTokenFilter = jwtTokenFilter;
-        this.corsConfigurationSource = corsConfigurationSource;
-    }
-    /**
-     * Setup default Security filter Chain
-     * @param http HttpSecurity for configuration
-     * @return Security Filter Chain to as default
-     * @throws Exception processing failed.
-     */
-    @Bean
-    public SecurityFilterChain defaultSecurityFilterChain(HttpSecurity http) throws Exception {
-        http.cors(httpSecurityCorsConfigurer ->
-                        httpSecurityCorsConfigurer.configurationSource(this.corsConfigurationSource))
-                .csrf(AbstractHttpConfigurer::disable)
-                .sessionManagement(httpSecuritySessionManagementConfigurer ->
-                        httpSecuritySessionManagementConfigurer.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .authorizeHttpRequests(authorize -> authorize
-                        .requestMatchers("/actuator/**", "/v3/api-docs/**", "/api-docs/**", "/swagger-ui/**")
-                        .permitAll()
-                        .anyRequest()
-                        .authenticated())
-                .exceptionHandling(httpSecurityExceptionHandlingConfigurer ->
-                        httpSecurityExceptionHandlingConfigurer.authenticationEntryPoint((request, response, ex) ->
-                                response.sendError(HttpServletResponse.SC_UNAUTHORIZED, ex.getMessage())))
-                .addFilterBefore(this.jwtTokenFilter, UsernamePasswordAuthenticationFilter.class);
-        return http.build();
-    }
-
-    /**
-     * Expose authentication manager bean
-     * @return the AuthenticationManager
-     * @throws Exception error with authentication manager.
-     */
-    @Bean
-    public AuthenticationManager authenticationManagerBean() throws Exception {
-        return this.configuration.getAuthenticationManager();
-    }
-}
