--- opencdx-predictor/src/main/java/cdx/opencdx/predictor/security/package-info.java
@@ -1,11 +1,0 @@
-/**
- * The purpose of this package is to encompass security-related configurations for Spring Security.
- * <p>
- * To facilitate easier identification and lookup, files related to Spring Security are logically grouped and separated within this package.
- * </p>
- *
- * @author Safe Health LLC.
- * @version 1.0
- * @since 1.0
- */
-package cdx.opencdx.predictor.security;
