--- opencdx-predictor/src/main/java/cdx/opencdx/predictor/service/impl/package-info.java
@@ -1,16 +1,0 @@
-/**
- * This package is dedicated to the concrete implementation of Spring Service Interfaces.
- * <p>
- * In the context of Spring Services, an interface typically represents the contract for a service. This package provides their concrete implementations.
- * In scenarios where multiple implementations for a single service are created, it's recommended to use the @Qualifier annotation for clear identification of each service instance.
- * For further details on fine-tuning annotation-based autowiring with qualifiers, please refer to the official Spring documentation:
- * <a href="https://docs.spring.io/spring-framework/docs/4.3.12.RELEASE/spring-framework-reference/htmlsingle/#beans-autowired-annotation-qualifiers">
- * Fine-tuning annotation-based autowiring with qualifiers.
- * </a>
- * </p>
- *
- * @author Safe Health LLC.
- * @version 1.0
- * @since 1.0
- */
-package cdx.opencdx.predictor.service.impl;
