--- opencdx-predictor/src/main/java/cdx/opencdx/predictor/service/package-info.java
@@ -1,12 +1,0 @@
-/**
- * This is the Service Interface package, designed specifically for defining Spring service interfaces.
- * <p>
- * The files housed within this package are typically of the Interface type. The corresponding implementations of these interfaces
- * are located in the adjoining 'impl' package.
- * </p>
- *
- * @author Safe Health LLC.
- * @version 1.0
- * @since 1.0
- */
-package cdx.opencdx.predictor.service;
