--- opencdx-predictor/src/main/resources/application.yml
@@ -1,23 +1,0 @@
-spring:
-  cloud:
-    config:
-      tls:
-        key-store: file:/certs/opencdx-keystore.p12
-        trust-store-type: PKCS12
-        key-store-type: PKCS12
-        key-password: opencdx
-        trust-store: file:/certs/opencdx-keystore.p12
-        trust-store-password: opencdx
-        enabled: 'true'
-        key-store-password: opencdx
-      uri: https://Config:8888
-  config:
-    import: configserver:https://Config:8888
-  application:
-    name: predictor
-  profiles:
-    active: managed,mongo,nats,predictor
-
-logging:
-  level:
-    cdx.opencdx: debug
