--- opencdx-predictor/src/main/resources/banner.txt
@@ -1,9 +1,0 @@
-
- ,-----.                        ,-----.,------.                  ,------.                  ,--.,--.        ,--.
-'  .-.  ' ,---.  ,---. ,--,--, '  .--./|  .-.  \,--.  ,--.,-----.|  .--. ',--.--. ,---.  ,-|  |`--' ,---.,-'  '-. ,---. ,--.--.
-|  | |  || .-. || .-. :|      \|  |    |  |  \  :\  `'  / '-----'|  '--' ||  .--'| .-. :' .-. |,--.| .--''-.  .-'| .-. ||  .--'
-'  '-'  '| '-' '\   --.|  ||  |'  '--'\|  '--'  //  /.  \        |  | --' |  |   \   --.\ `-' ||  |\ `--.  |  |  ' '-' '|  |
- `-----' |  |-'  `----'`--''--' `-----'`-------''--'  '--'       `--'     `--'    `----' `---' `--' `---'  `--'   `---' `--'
-         `--'
-${application.title} ${application.version}
-Powered by Spring Boot ${spring-boot.version}
