--- opencdx-predictor/src/test/java/cdx/opencdx/predictor/config/AppConfigTest.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2024 Safe Health Systems, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package cdx.opencdx.predictor.config;
-
-import cdx.opencdx.commons.service.impl.NoOpOpenCDXMessageServiceImpl;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-@ActiveProfiles("test")
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(
-        properties = "spring.cloud.config.enabled=false",
-        classes = {AppConfig.class, NoOpOpenCDXMessageServiceImpl.class})
-class AppConfigTest {
-
-    @Autowired
-    AppProperties appProperties;
-
-    @Test
-    void testFormat() {
-        AppConfig appConfig = new AppConfig();
-        Assertions.assertEquals("%s!", appConfig.format(this.appProperties));
-    }
-}
