--- opencdx-predictor/src/test/resources/application-test.yml
@@ -1,42 +1,0 @@
-# Properties for a local development environment go here.
-server:
-  port: 8080
-
-grpc:
-  port: 0
-
-management:
-  endpoint:
-    health:
-      enabled: true
-logging:
-  level:
-    root: TRACE
-    health:
-      safe: DEBUG
-spring:
-  application:
-    name: predictor
-  cloud:
-    config:
-      enabled: false
-#  datasource:
-#    url: jdbc:h2:mem:opencdx
-#    driverClassName: org.h2.Driver
-#    username: sa
-#    password: password
-#  jpa:
-#    database-platform: org.hibernate.dialect.H2Dialect
-#  h2:
-#    console:
-#      enabled: true
-  spring.data.mongodb:
-    uri: mongodb+srv://<username>:<password>@database:27017/opencdx/
-    username: root
-    password: root
-
-opencdx:
-  predictor:
-    format: "%s!"
-  client:
-    trustStore: ../certs/opencdx-clients.pem
