--- opencdx-predictor/src/test/resources/logback-spring.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <root level="DEBUG">
-        <appender-ref ref="CONSOLE"/>
-    </root>
-</configuration>
