+++ settings.gradle
@@ -26,7 +26,6 @@ include('opencdx-helloworld',
         'opencdx-iam',
         'opencdx-routine',	
         'opencdx-protector',
-        'opencdx-predictor',
         'opencdx-questionnaire',
         'opencdx-classification',
         'opencdx-classification-analyzer',
